Wait two mintues before writing `all.traj`
We have been getting a lot of errors where espresso_tools can't find
certain information in the QE output files, but when we go back to the
run directory, the output files are complete and espresso_tools runs
just fine. We think this is because the log files are not done writing
when espresso_tools tries to read them. So we now wait two minutes for
it to finish writing.