Throw specific error if job was cut short
Towards the goal of rerunning walltimed calculations for QE, let's start
throwing an assertion error when the output file says that we ran out of
time. And to make sure that espresso_tools can get to this point, let's
give it another 3 minutes (from 2 to 5 minutes) of overhead time to get
that part done.